Initial Commit of C# Console App Version
Has VERY little functionality with reading JSONS from a SPECIFIC file path.  Can currently only read character JSONS.  I have been messing around with these character JSONS to try to fit in everything that a pathfinder character would have.  Next is adding a list of worn items, a list of items in bags, and bulk.  This means that I need an Item class, which I need to work on as well.  This project is pretty fun, and it's nice to finally have something to do with c#.