Lots of stuff added.  Created some stuff for classes (does not work yet)
Edited Characters.JSON, still need to read pathfinder 2e documentation (should I try this project with a simpler system first?).

Added ClassData.JSON.  This is in the pathfinder classes folder, I'm not sure if I should have a different folder for homebrew.

Adding characters is broken again, but this time on purpose.  I can fix it tomorrow.  The main reason it is broken is because it does not currently add everything in the character class, so when the program tries to show everything, it fails.  I may need to add a check for certain things (skills mostly) to see if they even have a score/modifier before trying to print them to the console.